backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djoser==2.2.3
drf-extra-fields==3.4.0
django-filter==24.3
Pillow==10.4.0
psycopg2-binary==2.9.9

